siguiendo con el mando

+++ src/killergame/Controlled.java
@@ -13,7 +13,7 @@ public abstract class Controlled extends Alive {
      * <!-- begin-user-doc -->
      * <!--  end-user-doc  --> @generated
      */
-    public Controlled(KillerGame kGame, int posX, int posY, int width, int height, double velX, double velY) {
+    public Controlled(KillerGame kGame, double posX, double posY, int width, int height, double velX, double velY) {
         super(kGame, posX, posY, width, height, velX, velY);
     }
